sitemap upd1

+++ scripts/sitemap.sh
@@ -5,12 +5,16 @@ dest_file="${base_dir}/../sitemap.xml"
 bck_file="${base_dir}/../../backup/sitemap.xml"
 replacer="/./"
 [[ -f "${bck_file}" ]] && rm -f "${bck_file}"
-[[ -f "${dest_file}" ]] && cp "${dest_file}" "${bck_file}"
+
 cd "${base_dir}"
-echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" > "${dest_file}"
-echo "<urlset xmlns=\"http://www.sitemaps.org/schemas/sitemap/0.9\">" >> "${dest_file}"
-sed -i '$ d' "${dest_file}"
-for curr_file in $(find . -type f); do
+if [[ ! -f "${dest_file}" ]]; then
+  echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" > "${dest_file}"
+  echo "<urlset xmlns=\"http://www.sitemaps.org/schemas/sitemap/0.9\">" >> "${dest_file}"
+else 
+  cp "${dest_file}" "${bck_file}"
+  sed -i '$ d' "${dest_file}"
+fi
+for curr_file in $(find . -type f -not -path "./draft/*"); do
   mod=$(date -r ${curr_file} +%Y-%m-%d)
   name=${curr_file%.*}
   name=$(echo "${base_url}/${name}")
